iter5 -> refactoring Reorganized the structure of application
