+++ cmd/server/main.go
@@ -2,29 +2,15 @@ package main
 
 import (
 	"log"
-	"net/http"
 
-	h "github.com/sanek1/metrics-collector/internal/handlers"
-	rc "github.com/sanek1/metrics-collector/internal/routing"
-	s "github.com/sanek1/metrics-collector/internal/storage"
+	"github.com/sanek1/metrics-collector/internal/app"
 )
 
 func main() {
 	ParseFlags()
-	if err := RunServer(); err != nil {
+
+	application := app.New(Options.flagRunAddr)
+	if err := application.Run(); err != nil {
 		log.Fatal(err)
 	}
-}
-
-func RunServer() error {
-
-	memStorage := s.NewMemStorage()
-	metricStorage := h.MetricStorage{
-		Storage: memStorage,
-	}
-
-	r := rc.InitRouting(metricStorage)
-	log.Println("Server start on ", Options.flagRunAddr)
-	log.Fatal(http.ListenAndServe(Options.flagRunAddr, r))
-	return nil
 }
