+++ internal/handlers/metric_storage.go
@@ -1,0 +1,9 @@
+package handlers
+
+import (
+	"github.com/sanek1/metrics-collector/internal/storage"
+)
+
+type MetricStorage struct {
+	Storage storage.Storage
+}
