+++ internal/routing/routing.go
@@ -12,14 +12,18 @@ import (
 func InitRouting(ms h.MetricStorage) http.Handler {
 	r := chi.NewRouter()
 	r.Use(middleware.Logger)
+
+	// get
 	r.Get("/*", http.HandlerFunc(ms.MainPageHandler))
 	r.Get("/{value}/{type}/*", http.HandlerFunc(ms.GetMetricsByNameHandler))
+
+	// post
+	r.Post("/*", v.Validation(http.HandlerFunc(h.NotImplementedHandler)))
 	r.Route("/update", func(r chi.Router) {
 		r.Post("/*", v.Validation(http.HandlerFunc(h.BadRequestHandler)))
 		r.Post("/gauge/*", v.Validation(http.HandlerFunc(ms.GaugeHandler)))
 		r.Post("/counter/*", v.Validation(http.HandlerFunc(ms.CounterHandler)))
 	})
-	r.Post("/*", v.Validation(http.HandlerFunc(h.NotImplementedHandler)))
-	r.Get("/*", http.HandlerFunc(h.NotImplementedHandler))
+
 	return r
 }
